app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
motor==3.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas>=2.0.0
scikit-learn==1.3.2
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.4.2
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
pydantic[email]
email-validator==2.1.0.post1 